Move the http-specific fetch code to http_loader.
